# Conflicts:
#	ccv/chains/evm/deployment/go.mod
#	ccv/chains/evm/deployment/go.sum
#	integration-tests/go.mod

+++ chains/evm/contracts/test/mocks/MockUSDCTokenMessenger.sol
@@ -77,7 +77,7 @@ contract MockUSDCTokenMessenger is ITokenMessenger {
     bytes32 mintRecipient,
     address burnToken,
     bytes32 destinationCaller,
-    uint32 maxFee,
+    uint256 maxFee,
     uint32 minFinalityThreshold,
     bytes calldata hookData
   ) external {
